moved ntt code into polynomial
