+++ libcrux-ml-kem/src/lib.rs
@@ -16,7 +16,6 @@ mod constant_time_ops;
 mod hash_functions;
 mod ind_cpa;
 mod matrix;
-mod ntt;
 mod polynomial;
 mod sampling;
 mod serialize;
